# swagger.yaml
openapi: 3.0.0
info:
  title: Node Tracker API
  version: 1.0.0
  description: REST API for a private BitTorrent tracker with user management, torrents, and IP bans.
  contact:
    name: API Support
    email: support@nodetracker.com
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          description: Unique user ID
        username:
          type: string
          description: Username
        email:
          type: string
          format: email
          description: User email
        role:
          type: string
          enum: [USER, MODERATOR, ADMIN]
          description: User role
        banned:
          type: boolean
          description: Whether the user is banned
        emailVerified:
          type: boolean
          description: Whether the email is verified
        created:
          type: string
          format: date-time
          description: Creation date
        uploaded:
          type: integer
          description: Total bytes uploaded by the user
        downloaded:
          type: integer
          description: Total bytes downloaded by the user
        seedtime:
          type: integer
          description: Total seed time in seconds
        ratio:
          type: number
          format: float
          description: Upload/download ratio (uploaded/downloaded)
...
    UserBan:
      type: object
      properties:
        id:
          type: integer
          description: Unique ban ID
        userId:
          type: integer
          description: Banned user ID
        reason:
          type: string
          description: Ban reason
        bannedBy:
          type: string
          description: Username of admin/moderator who applied the ban
        bannedAt:
          type: string
          format: date-time
          description: Ban date and time
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: Expiration date (null for permanent ban)
        active:
          type: boolean
          description: Whether the ban is active
        user:
          type: object
          properties:
            id:
              type: integer
            username:
              type: string
            email:
              type: string
              format: email
...
    UserBanRequest:
      type: object
      required: [userId, reason]
      properties:
        userId:
          type: integer
          minimum: 1
          description: ID of user to ban
        reason:
          type: string
          minLength: 5
          maxLength: 500
          description: Ban reason
        expiresAt:
          type: string
          format: date-time
          description: Expiration date (optional, null for permanent)
    QuickBanRequest:
      type: object
      required: [userId, reason]
      properties:
        userId:
          type: integer
          minimum: 1
          description: ID of user to ban
        reason:
          type: string
          minLength: 5
          maxLength: 500
          description: Ban reason
    CustomBanRequest:
      type: object
      required: [userId, reason, days]
      properties:
        userId:
          type: integer
          minimum: 1
          description: ID of user to ban
        reason:
          type: string
          minLength: 5
          maxLength: 500
          description: Ban reason
        days:
          type: integer
          minimum: 1
          maximum: 365
          description: Number of ban days
...
    CleanupResult:
      type: object
      properties:
        message:
          type: string
          description: Result message
        cleaned:
          type: integer
          description: Number of bans cleaned
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
security:
  - bearerAuth: []
...
paths:
...
  /api/users:
    get:
      summary: List all users (admin only)
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Results limit per page
      responses:
        200:
          description: User list with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        403:
          description: Access denied
  /api/users/{id}:
    get:
      summary: Get specific user information
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        200:
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        404:
          description: User not found
...
  /api/user-bans:
    get:
      summary: List all user bans
      tags: [UserBans]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Results limit per page
        - in: query
          name: active
          schema:
            type: boolean
          description: Filter by active bans
        - in: query
          name: userId
          schema:
            type: integer
          description: Filter by user ID
        - in: query
          name: bannedBy
          schema:
            type: string
          description: Filter by who applied the ban
      responses:
        200:
          description: Ban list with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  userBans:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserBan'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      summary: Create a new user ban
      tags: [UserBans]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserBanRequest'
      responses:
        201:
          description: User ban created successfully
  /api/user-bans/quick/7-days:
    post:
      summary: Ban user for 7 days
      tags: [UserBans]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuickBanRequest'
      responses:
        201:
          description: User banned for 7 days
  /api/user-bans/quick/15-days:
    post:
      summary: Ban user for 15 days
      tags: [UserBans]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuickBanRequest'
      responses:
        201:
          description: User banned for 15 days
  /api/user-bans/quick/30-days:
    post:
      summary: Ban user for 30 days
      tags: [UserBans]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuickBanRequest'
      responses:
        201:
          description: User banned for 30 days
  /api/user-bans/quick/permanent:
    post:
      summary: Ban user permanently
      tags: [UserBans]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuickBanRequest'
      responses:
        201:
          description: User banned permanently
  /api/user-bans/custom:
    post:
      summary: Ban user for specific days
      tags: [UserBans]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomBanRequest'
      responses:
        201:
          description: User banned for specific period
  /api/user-bans/{id}:
    get:
      summary: Get specific ban information
      tags: [UserBans]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Ban ID
      responses:
        200:
          description: Ban information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBan'

    delete:
      summary: Deactivate a user ban
      tags: [UserBans]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Ban ID
      responses:
        200:
          description: Ban deactivated successfully
  /api/user-bans/user/{userId}:
    get:
      summary: Get bans for a specific user
      tags: [UserBans]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        200:
          description: User ban list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserBan'
...
  /api/user-bans/cleanup:
    post:
      summary: Clean up expired bans
      tags: [UserBans]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Cleanup completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CleanupResult'
...